axconfig_gen_macros
Macro include_configs

include_configs!() { /* proc-macro */ }

Includes a TOML format config file and expands it into Rust code.

The given path should be an absolute path or a path relative to your project’s Cargo.toml.

Example

See the crate-level documentation.